Update copy between map to copy also dart_info; update also isomorphism function; start to update test (nyf)
