+++ Linear_cell_complex/test/Linear_cell_complex/CMakeLists.txt
@@ -16,10 +16,14 @@ if ( CGAL_FOUND )
 
   include_directories (BEFORE "../../include")
 
+  file(GLOB hfiles RELATIVE ${CMAKE_CURRENT_SOURCE_DIR}
+        ${CMAKE_CURRENT_SOURCE_DIR}/*.h)
+
   # create a target per cppfile
-  file(GLOB cppfiles RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_SOURCE_DIR}/*.cpp)
+  file(GLOB cppfiles RELATIVE ${CMAKE_CURRENT_SOURCE_DIR}
+        ${CMAKE_CURRENT_SOURCE_DIR}/*.cpp)
   foreach(cppfile ${cppfiles})
-    create_single_source_cgal_program( "${cppfile}" )
+    create_single_source_cgal_program( "${cppfile}" "${hfiles}" )
   endforeach()
 
 else()
