+++ Linear_cell_complex/test/Linear_cell_complex/Linear_cell_complex_4_test.h
@@ -20,7 +20,7 @@
 #ifndef CGAL_LCC_4_TEST_H
 #define CGAL_LCC_4_TEST_H
 
-#include <CGAL/Linear_cell_complex.h>
+#include <CGAL/Linear_cell_complex_for_combinatorial_map.h>
 #include <CGAL/Combinatorial_map_operations.h>
 #include "Linear_cell_complex_2_test.h"
